--- test/integration/app/search-packages.test.ts
@@ -1,96 +1,0 @@
-import { searchPackagesUsing } from "../../../src/app/search-packages";
-import { DomainName } from "../../../src/domain/domain-name";
-import { Registry } from "../../../src/domain/registry";
-import { SemanticVersion } from "../../../src/domain/semantic-version";
-import {
-  AllPackuments,
-  GetAllRegistryPackuments,
-} from "../../../src/io/all-packuments-io";
-import { SearchedPackument, SearchRegistry } from "../../../src/io/npm-search";
-import { noopLogger } from "../../../src/logging";
-import { partialApply } from "../../../src/utils/fp-utils";
-import { exampleRegistryUrl } from "../../common/data-registry";
-import { mockFunctionOfType } from "./func.mock";
-
-describe("search packages", () => {
-  const exampleRegistry: Registry = {
-    url: exampleRegistryUrl,
-    auth: null,
-  };
-
-  const exampleKeyword = "package-a";
-
-  const exampleSearchResult: SearchedPackument = {
-    name: DomainName.parse("com.example.package-a"),
-    versions: { [SemanticVersion.parse("1.0.0")]: "latest" },
-    description: "A demo package",
-    date: new Date(2019, 9, 2, 3, 2, 38),
-    "dist-tags": { latest: SemanticVersion.parse("1.0.0") },
-  };
-
-  const exampleAllPackumentsResult = {
-    _updated: 99999,
-    [exampleSearchResult.name]: exampleSearchResult,
-  } as AllPackuments;
-
-  function makeDependencies() {
-    const searchRegistry = mockFunctionOfType<SearchRegistry>();
-    searchRegistry.mockResolvedValue([exampleSearchResult]);
-
-    const fetchAllPackument = mockFunctionOfType<GetAllRegistryPackuments>();
-    fetchAllPackument.mockResolvedValue(exampleAllPackumentsResult);
-
-    const searchPackages = partialApply(
-      searchPackagesUsing,
-      searchRegistry,
-      fetchAllPackument,
-      noopLogger
-    );
-    return { searchPackages, searchRegistry, fetchAllPackument } as const;
-  }
-
-  it("should search using search api first", async () => {
-    const { searchPackages } = makeDependencies();
-
-    const actual = await searchPackages(exampleRegistry, exampleKeyword);
-
-    expect(actual).toEqual([exampleSearchResult]);
-  });
-
-  it("should not notify of fallback when using search api", async () => {
-    const { searchPackages } = makeDependencies();
-
-    const fallback = jest.fn();
-    await searchPackages(exampleRegistry, exampleKeyword, fallback);
-
-    expect(fallback).not.toHaveBeenCalled();
-  });
-
-  it("should search using old search if search api is not available", async () => {
-    const { searchPackages, searchRegistry } = makeDependencies();
-    searchRegistry.mockRejectedValue(new Error());
-
-    const actual = await searchPackages(exampleRegistry, exampleKeyword);
-
-    expect(actual).toEqual([exampleSearchResult]);
-  });
-
-  it("should notify of using old search", async () => {
-    const { searchPackages, searchRegistry } = makeDependencies();
-    searchRegistry.mockRejectedValue(new Error());
-
-    const fallback = jest.fn();
-    await searchPackages(exampleRegistry, exampleKeyword, fallback);
-
-    expect(fallback).toHaveBeenCalled();
-  });
-
-  it("should not find packages not matching the keyword using old search", async () => {
-    const { searchPackages, searchRegistry } = makeDependencies();
-    searchRegistry.mockRejectedValue(new Error());
-
-    const actual = await searchPackages(exampleRegistry, "some other keyword");
-
-    expect(actual).toEqual([]);
-  });
-});
