+++ test/integration/io/upm-config-io.test.ts
@@ -3,7 +3,7 @@ import { ReadTextFile } from "../../../src/io/text-file-io";
 import { loadUpmConfigUsing } from "../../../src/io/upm-config-io";
 import { partialApply } from "../../../src/utils/fp-utils";
 import { exampleRegistryUrl } from "../../common/data-registry";
-import { mockFunctionOfType } from "../app/func.mock";
+import { mockFunctionOfType } from "../func.mock";
 
 jest.mock("../../../src/utils/env-util");
