+++ test/unit/app/get-registry-auth.test.ts
