+++ test/unit/app/put-registry-auth.test.ts
